Fix Speed animations and modify Speed cards transition

+++ Source/Guis/GuiTextBox.h
@@ -30,7 +30,7 @@ namespace ClassicLauncher
         int mDesiredWidth;
         float mOffset;
         bool mToLeft;
-        int mDelay;
+        float mDelay;
         Vector2 mMensuredText;
         float mSpeed;
         float mMaxDelay;
